refactor after_prepare hook
