--- scripts/android/after_prepare.js
@@ -1,39 +1,0 @@
-#!/usr/bin/env node
-
-'use strict';
-
-/**
- * This hook makes sure projects using [cordova-plugin-firebase](https://github.com/arnesson/cordova-plugin-firebase)
- * will build properly and have the required key files copied to the proper destinations when the app is build on Ionic Cloud using the package command.
- * Credits: https://github.com/arnesson.
- */
-var fs = require('fs');
-var path = require('path');
-var utilities = require("../lib/utilities");
-
-var config = fs.readFileSync('config.xml').toString();
-var name = utilities.getValue(config, 'name');
-
-var ANDROID_DIR = 'platforms/android';
-
-var PLATFORM = {
-  ANDROID: {
-    dest: [
-      ANDROID_DIR + '/google-services.json',
-      ANDROID_DIR + '/app/google-services.json'
-    ],
-    src: [
-      'google-services.json',
-      ANDROID_DIR + '/assets/www/google-services.json',
-      'www/google-services.json',
-      ANDROID_DIR + '/app/src/main/google-services.json'
-    ],
-  }
-};
-
-module.exports = function (context) {
-  if (utilities.directoryExists(ANDROID_DIR)) {
-    console.log('Preparing Firebase on Android');
-    utilities.copyKey(PLATFORM.ANDROID);
-  }
-};
