+++ scripts/lib/utilities.js
@@ -4,20 +4,7 @@
 var fs = require('fs');
 var path = require("path");
 
-fs.ensureDirSync = function (dir) {
-  if (!fs.existsSync(dir)) {
-    dir.split(path.sep).reduce(function (currentPath, folder) {
-      currentPath += folder + path.sep;
-      if (!fs.existsSync(currentPath)) {
-        fs.mkdirSync(currentPath);
-      }
-      return currentPath;
-    }, '');
-  }
-};
-
 module.exports = {
-
   /**
      * Used to get the name of the application as defined in the config.xml.
      *
